Dev: Code cleanup and reorganization
