+++ quantum_launcher/src/menu_renderer/launch.rs
@@ -11,8 +11,8 @@ use crate::menu_renderer::{
     view_info_message,
 };
 use crate::state::{
-    GameLogMessage, InstanceNotes, LaunchModal, LauncherSettingsTab, MainMenuMessage, NotesMessage,
-    ShortcutMessage, SidebarMessage, SidebarScroll, WindowMessage,
+    GameLogMessage, InstanceNotes, LaunchMessage, LaunchModal, LauncherSettingsTab,
+    MainMenuMessage, NotesMessage, ShortcutMessage, SidebarMessage, SidebarScroll, WindowMessage,
 };
 use crate::{
     icons,
@@ -483,7 +483,7 @@ impl Launcher {
         } else if self.processes.contains_key(selected) {
             tooltip(
                 button_with_icon(icons::play(), "Kill", 16)
-                    .on_press(Message::LaunchKill)
+                    .on_press(LaunchMessage::Kill.into())
                     .width(98),
                 shortcut_ctrl("Backspace"),
                 Position::Bottom,
@@ -496,7 +496,7 @@ impl Launcher {
             )
         } else {
             tooltip(
-                play_button.on_press(Message::LaunchStart),
+                play_button.on_press(LaunchMessage::Start.into()),
                 shortcut_ctrl("Enter"),
                 Position::Bottom,
             )
@@ -521,7 +521,7 @@ impl Launcher {
             tooltip(
                 button_with_icon(icons::play(), "Stop", 16)
                     .width(98)
-                    .on_press(Message::LaunchKill),
+                    .on_press(LaunchMessage::Kill.into()),
                 shortcut_ctrl("Escape"),
                 Position::Bottom,
             )
@@ -535,7 +535,7 @@ impl Launcher {
             tooltip(
                 button_with_icon(icons::play(), "Start", 16)
                     .width(98)
-                    .on_press(Message::LaunchStart),
+                    .on_press(LaunchMessage::Start.into()),
                 "By starting the server, you agree to the EULA",
                 Position::Bottom,
             )
