+++ quantum_launcher/src/message_handler/iced_event.rs
@@ -1,11 +1,11 @@
 use crate::message_handler::arrow_keys::InstSelectOperation;
 use crate::message_update::MSG_RESIZE;
 use crate::state::{
-    AutoSaveKind, CreateInstanceMessage, InfoMessage, LaunchTab, Launcher, LauncherSettingsMessage,
-    LauncherSettingsTab, MainMenuMessage, ManageModsMessage, MenuCreateInstance,
-    MenuCreateInstanceChoosing, MenuEditMods, MenuEditPresets, MenuExportInstance,
-    MenuInstallFabric, MenuInstallOptifine, MenuInstallPaper, MenuLoginAlternate, MenuLoginMS,
-    MenuRecommendedMods, MenuWelcome, Message, State,
+    AutoSaveKind, CreateInstanceMessage, InfoMessage, LaunchMessage, LaunchTab, Launcher,
+    LauncherSettingsMessage, LauncherSettingsTab, MainMenuMessage, ManageModsMessage,
+    MenuCreateInstance, MenuCreateInstanceChoosing, MenuEditMods, MenuEditPresets,
+    MenuExportInstance, MenuInstallFabric, MenuInstallOptifine, MenuInstallPaper,
+    MenuLoginAlternate, MenuLoginMS, MenuRecommendedMods, MenuWelcome, Message, State,
 };
 use iced::{
     Task,
@@ -196,11 +196,11 @@ impl Launcher {
                 return self.select_instance_recursive(InstSelectOperation::Down);
             } else if let Key::Named(Named::Enter) = key {
                 if modifiers.command() {
-                    return self.launch_start();
+                    return Task::done(LaunchMessage::Start.into());
                 }
             } else if let Key::Named(Named::Backspace) = key {
                 if modifiers.command() {
-                    return Task::done(Message::LaunchKill);
+                    return Task::done(LaunchMessage::Kill.into());
                 }
             }
         } else if let State::Create(MenuCreateInstance::Choosing(MenuCreateInstanceChoosing {
