+++ quantum_launcher/src/message_handler/mod.rs
@@ -1,35 +1,25 @@
 use crate::{
     Launcher, Message,
-    config::AfterLaunchBehavior,
     menu_renderer::back_to_launch_screen,
     state::{
-        AutoSaveKind, EditPresetsMessage, GameProcess, InfoMessage, LaunchTab, LogState,
-        ManageModsMessage, MenuCreateInstance, MenuCreateInstanceChoosing, MenuEditInstance,
-        MenuEditMods, MenuInstallForge, MenuInstallOptifine, MenuLaunch, OFFLINE_ACCOUNT_NAME,
-        ProgressBar, SelectedState, State,
+        AutoSaveKind, EditPresetsMessage, InfoMessage, LaunchTab, LogState, ManageModsMessage,
+        MenuEditMods, MenuInstallForge, MenuInstallOptifine, ProgressBar, SelectedState, State,
     },
     tick::sort_dependencies,
 };
 use iced::{Task, futures::executor::block_on, widget::scrollable::AbsoluteOffset};
 use ql_core::{
-    GenericProgress, Instance, InstanceKind, IntoIoError, IntoJsonError, IntoStringError,
-    JsonFileError, LaunchedProcess, err,
+    GenericProgress, Instance, IntoIoError, IntoStringError, JsonFileError, err,
     file_utils::exists,
-    info,
     json::{VersionDetails, instance_config::InstanceConfigJson},
-    pt,
-    read_log::{Diagnostic, ReadError},
 };
-use ql_instances::auth::AccountData;
 use ql_mod_manager::{loaders, store::ModIndex};
 use std::{
     collections::HashSet,
     ffi::OsStr,
     path::{Path, PathBuf},
-    process::ExitStatus,
     sync::mpsc::{Receiver, Sender},
 };
-use tokio::io::AsyncWriteExt;
 
 pub const SIDEBAR_LIMIT_RIGHT: f32 = 140.0;
 pub const SIDEBAR_LIMIT_LEFT: f32 = 135.0;
@@ -74,110 +64,6 @@ impl Launcher {
         }
     }
 
-    pub fn launch_game(&mut self, account_data: Option<AccountData>) -> Task<Message> {
-        let username = if let Some(account_data) = &account_data {
-            // Logged in account
-            account_data.nice_username.clone()
-        } else {
-            // Offline username
-            self.config.username.clone()
-        };
-
-        let (sender, receiver) = std::sync::mpsc::channel();
-        self.java_recv = Some(ProgressBar::with_recv(receiver));
-
-        let global_settings = self.config.global_settings.clone();
-        let extra_java_args = self.config.extra_java_args.clone().unwrap_or_default();
-
-        let instance_name = self.instance().name.clone();
-        Task::perform(
-            async move {
-                ql_instances::launch(
-                    instance_name,
-                    username,
-                    Some(sender),
-                    account_data,
-                    global_settings,
-                    extra_java_args,
-                )
-                .await
-                .strerr()
-            },
-            Message::LaunchEnd,
-        )
-    }
-
-    pub fn finish_launching(&mut self, result: Result<LaunchedProcess, String>) -> Task<Message> {
-        self.java_recv = None;
-        self.is_launching_game = false;
-        match result {
-            Ok(child) => {
-                let selected_instance = child.instance.clone();
-
-                let server_input = block_on(child.child.lock())
-                    .stdin
-                    .take()
-                    .map(|n| (n, false));
-
-                let (sender, receiver) = std::sync::mpsc::channel();
-                self.processes.insert(
-                    selected_instance.clone(),
-                    GameProcess {
-                        child: child.clone(),
-                        receiver: Some(receiver),
-                        server_input,
-                    },
-                );
-
-                let mut censors = Vec::new();
-                for account in self.accounts.values() {
-                    if let Some(token) = &account.access_token {
-                        censors.push(token.clone());
-                    }
-                }
-
-                let version_presence_task = self.rpc_game_update(selected_instance.clone(), false);
-
-                let log_task = Task::perform(
-                    async move {
-                        let result = child.read_logs(censors, Some(sender)).await;
-                        let default_output = Ok((ExitStatus::default(), selected_instance, None));
-
-                        match result {
-                            Some(Err(ReadError::Io(io)))
-                                if io.kind() == std::io::ErrorKind::InvalidData =>
-                            {
-                                err!("Minecraft log contains invalid unicode! Stopping logs");
-                                pt!("The game will continue to run");
-                                default_output
-                            }
-                            Some(result) => result.strerr(),
-                            None => default_output,
-                        }
-                    },
-                    Message::LaunchGameExited,
-                );
-
-                match self.config.c_after_launch_behavior() {
-                    AfterLaunchBehavior::DoNothing => {}
-                    AfterLaunchBehavior::CloseLauncher => {
-                        ql_core::logger_finish();
-                        std::process::exit(0);
-                    }
-                    AfterLaunchBehavior::MinimizeLauncher => {
-                        let minimize_task = iced::window::get_latest()
-                            .and_then(|id| iced::window::minimize(id, true));
-                        return Task::batch([log_task, minimize_task, version_presence_task]);
-                    }
-                }
-
-                return Task::batch([log_task, version_presence_task]);
-            }
-            Err(err) => self.set_error(err),
-        }
-        Task::none()
-    }
-
     pub fn delete_instance_confirm(&mut self) -> Task<Message> {
         let State::ConfirmAction { .. } = &self.state else {
             return Task::none();
@@ -205,36 +91,6 @@ impl Launcher {
         p.selected_instance = None;
         p.selected_instance_kind = None;
         self.autosave.remove(&AutoSaveKind::LauncherConfig);
-    }
-
-    pub fn load_edit_instance_inner(
-        edit_instance: &mut Option<MenuEditInstance>,
-        selected_instance: &Instance,
-    ) -> Result<(), JsonFileError> {
-        let config_path = selected_instance.get_instance_path().join("config.json");
-
-        let config_json = std::fs::read_to_string(&config_path).path(config_path)?;
-        let config_json: InstanceConfigJson =
-            serde_json::from_str(&config_json).json(config_json)?;
-
-        let slider_value = f32::log2(config_json.ram_in_mb as f32);
-        let memory_mb = config_json.ram_in_mb;
-
-        // Use this to check for performance impact
-        // std::thread::sleep(std::time::Duration::from_millis(500));
-
-        *edit_instance = Some(MenuEditInstance {
-            main_class_mode: config_json.get_main_class_mode(),
-            config: config_json,
-            slider_value,
-            instance_name: selected_instance.name.to_string(),
-            old_instance_name: selected_instance.name.clone(),
-            slider_text: format_memory(memory_mb),
-            memory_input: memory_mb.to_string(),
-            is_editing_name: false,
-            arg_split_by_space: true,
-        });
-        Ok(())
     }
 
     pub fn go_to_edit_mods_menu(&mut self, msg: Option<InfoMessage>) -> Task<Message> {
@@ -295,87 +151,6 @@ impl Launcher {
         }
     }
 
-    pub fn set_game_exited(
-        &mut self,
-        status: ExitStatus,
-        instance: &Instance,
-        diagnostic: Option<Diagnostic>,
-    ) -> Task<Message> {
-        let kind = if instance.is_server() {
-            "Server"
-        } else {
-            "Game"
-        };
-        info!("Game exited ({status})");
-
-        let log_state = if let State::Launch(MenuLaunch {
-            message, log_state, ..
-        }) = &mut self.state
-        {
-            let has_crashed = !status.success();
-            if has_crashed {
-                let mut msg = format!("{kind} crashed! ({status})\nCheck \"Logs\" for more info");
-                if let Some(diag) = diagnostic {
-                    msg.push_str("\n\n");
-                    msg.push_str(&diag.to_string());
-                }
-                *message = Some(InfoMessage::error(msg));
-            }
-            if let Some(log) = self.logs.get_mut(instance) {
-                log.has_crashed = has_crashed;
-            }
-            log_state
-        } else {
-            &mut None
-        };
-
-        if let Some(process) = self.processes.remove(instance) {
-            Self::read_game_logs(
-                &process,
-                instance,
-                &mut self.logs,
-                log_state,
-                self.selected_instance.as_ref(),
-            );
-        }
-
-        self.rpc_game_update(instance.clone(), true)
-    }
-
-    pub fn update_mods(&mut self) -> Task<Message> {
-        if let State::EditMods(menu) = &mut self.state {
-            let updates = menu
-                .available_updates
-                .clone()
-                .into_iter()
-                .map(|(id, version, _)| (id, version))
-                .collect();
-            let write_changelog = self.config.c_persistent().write_mod_update_changelog;
-            let (sender, receiver) = std::sync::mpsc::channel();
-            menu.mod_update_progress = Some(ProgressBar::with_recv_and_msg(
-                receiver,
-                "Deleting Mods".to_owned(),
-            ));
-            let selected_instance = self.selected_instance.clone().unwrap();
-            Task::perform(
-                ql_mod_manager::store::apply_updates(
-                    selected_instance,
-                    updates,
-                    Some(sender),
-                    write_changelog,
-                ),
-                move |n| {
-                    ManageModsMessage::UpdatePerformDone(
-                        n.strerr().map(|res| (res, write_changelog)),
-                    )
-                    .into()
-                },
-            )
-        } else {
-            Task::none()
-        }
-    }
-
     pub fn install_forge(&mut self, kind: ForgeKind) -> Task<Message> {
         let (f_sender, f_receiver) = std::sync::mpsc::channel();
         let (j_sender, j_receiver): (Sender<GenericProgress>, Receiver<GenericProgress>) =
@@ -424,24 +199,6 @@ impl Launcher {
             is_java_getting_installed: false,
         });
         command
-    }
-
-    pub fn selected_kind(&self) -> Option<InstanceKind> {
-        self.selected_instance.as_ref().map(|n| n.kind).or(
-            if let State::Create(MenuCreateInstance::Choosing(MenuCreateInstanceChoosing {
-                kind,
-                ..
-            })) = &self.state
-            {
-                Some(*kind)
-            } else {
-                None
-            },
-        )
-    }
-
-    pub fn get_selected_dot_minecraft_dir(&self) -> Option<PathBuf> {
-        Some(self.selected_instance.as_ref()?.get_dot_minecraft_path())
     }
 
     fn load_modpack_from_path(&mut self, path: PathBuf) -> Task<Message> {
@@ -547,37 +304,6 @@ impl Launcher {
         }
     }
 
-    pub fn kill_selected_instance(&mut self) -> Task<Message> {
-        let Some(instance) = &self.selected_instance else {
-            return Task::none();
-        };
-        match instance.kind {
-            InstanceKind::Client => {
-                if let Some(process) = self.processes.remove(instance) {
-                    let mut child = block_on(process.child.child.lock());
-                    _ = child.start_kill();
-                }
-            }
-            InstanceKind::Server => {
-                if let Some(GameProcess {
-                    server_input: Some((stdin, has_issued_stop_command)),
-                    child,
-                    ..
-                }) = self.processes.get_mut(instance)
-                {
-                    *has_issued_stop_command = true;
-                    if child.is_classic_server {
-                        _ = block_on(child.child.lock()).start_kill();
-                    } else {
-                        let future = stdin.write_all("stop\n".as_bytes());
-                        _ = block_on(future);
-                    }
-                }
-            }
-        }
-        Task::none()
-    }
-
     pub fn go_to_delete_instance_menu(&mut self) {
         let instance = self.instance();
         self.state = State::ConfirmAction {
@@ -594,49 +320,6 @@ impl Launcher {
             yes: Message::DeleteInstance,
             no: back_to_launch_screen(None),
         };
-    }
-
-    pub fn launch_start(&mut self) -> Task<Message> {
-        let Some(selected_instance) = &self.selected_instance else {
-            return Task::none();
-        };
-        if self.processes.contains_key(selected_instance) {
-            return Task::none();
-        }
-        self.logs.remove(selected_instance);
-
-        match selected_instance.kind {
-            InstanceKind::Client => {
-                if self.account_selected == OFFLINE_ACCOUNT_NAME
-                    && (self.config.username.is_empty() || self.config.username.contains(' '))
-                {
-                    return Task::none();
-                }
-
-                self.is_launching_game = true;
-                let account_data = self.get_selected_account_data();
-                // If the user is loading an existing login from disk
-                // then first refresh the tokens
-                if let Some(account) = &account_data {
-                    if account.access_token.is_none() || account.needs_refresh {
-                        return self.account_refresh(account);
-                    }
-                }
-                // Or, if the account is already refreshed/freshly added,
-                // directly launch the game
-                self.launch_game(account_data)
-            }
-            InstanceKind::Server => {
-                let (sender, receiver) = std::sync::mpsc::channel();
-                self.java_recv = Some(ProgressBar::with_recv(receiver));
-
-                let server = selected_instance.name.clone();
-                Task::perform(
-                    async move { ql_servers::run(server, Some(sender)).await.strerr() },
-                    Message::LaunchEnd,
-                )
-            }
-        }
     }
 }
 
@@ -667,16 +350,6 @@ pub async fn get_locally_installed_mods(
         }
     }
     set
-}
-
-pub fn format_memory(memory_bytes: usize) -> String {
-    const MB_TO_GB: usize = 1024;
-
-    if memory_bytes >= MB_TO_GB {
-        format!("{:.2} GB", memory_bytes as f64 / MB_TO_GB as f64)
-    } else {
-        format!("{memory_bytes} MB")
-    }
 }
 
 #[derive(Debug, Clone, Copy)]
