+++ quantum_launcher/src/state/message.rs
@@ -320,25 +320,25 @@ pub enum RpcMessage {
 
 #[derive(Debug, Clone)]
 pub enum RpcInnerMessage {
-    TopTextChanged(String),
-    TopTextURLChanged(String),
-    BottomTextChanged(String),
-    BottomTextURLChanged(String),
+    TopText(String),
+    TopTextURL(String),
+    BottomText(String),
+    BottomTextURL(String),
 }
 
 impl RpcText {
     pub fn apply(&mut self, msg: RpcInnerMessage) {
         match msg {
-            RpcInnerMessage::TopTextChanged(text) => {
+            RpcInnerMessage::TopText(text) => {
                 self.top_text = (!text.is_empty()).then_some(text);
             }
-            RpcInnerMessage::TopTextURLChanged(text) => {
+            RpcInnerMessage::TopTextURL(text) => {
                 self.top_text_url = (!text.is_empty()).then_some(text);
             }
-            RpcInnerMessage::BottomTextChanged(text) => {
+            RpcInnerMessage::BottomText(text) => {
                 self.bottom_text = (!text.is_empty()).then_some(text);
             }
-            RpcInnerMessage::BottomTextURLChanged(text) => {
+            RpcInnerMessage::BottomTextURL(text) => {
                 self.bottom_text_url = (!text.is_empty()).then_some(text);
             }
         }
@@ -459,6 +459,14 @@ pub enum ModDescriptionMessage {
 }
 
 #[derive(Debug, Clone)]
+pub enum LaunchMessage {
+    Start,
+    End(Res<LaunchedProcess>),
+    Kill,
+    GameExited(Res<(ExitStatus, Instance, Option<Diagnostic>)>),
+}
+
+#[derive(Debug, Clone)]
 pub enum Message {
     Nothing,
     Error(String),
@@ -468,6 +476,7 @@ pub enum Message {
     WelcomeContinueToTheme,
     WelcomeContinueToAuth,
 
+    Launch(LaunchMessage),
     Account(AccountMessage),
     CreateInstance(CreateInstanceMessage),
     EditInstance(EditInstanceMessage),
@@ -476,7 +485,6 @@ pub enum Message {
     GameLog(GameLogMessage),
     Window(WindowMessage),
     Shortcut(ShortcutMessage),
-
     ManageMods(ManageModsMessage),
     ManageJarMods(ManageJarModsMessage),
     InstallMods(InstallModsMessage),
@@ -493,10 +501,6 @@ pub enum Message {
         message: Option<InfoMessage>,
         clear_selection: bool,
     },
-    LaunchStart,
-    LaunchEnd(Res<LaunchedProcess>),
-    LaunchKill,
-    LaunchGameExited(Res<(ExitStatus, Instance, Option<Diagnostic>)>),
 
     DeleteInstanceMenu,
     DeleteInstance,
@@ -562,6 +566,7 @@ macro_rules! from_m {
     };
 }
 
+from_m!(Launch, LaunchMessage);
 from_m!(MainMenu, MainMenuMessage);
 from_m!(Sidebar, SidebarMessage);
 from_m!(ManageMods, ManageModsMessage);
